A whole bunch of modifications.
With the graph schedule function, now run:

>pcl_kinfu_largeScale_release.exe -r -ic -sd 10 -oni 10-24-fix\20121024-fix.oni --fragment 25 --fragment_start 0 --use_rgbdslam 20121024-151428_rgbdslam_20121120-180556.log --fragment_registration 10-24-fix\20121128-231315.graph

For each pair, it resets the volume and apply kinfu between two fragments.

Or we could change the processGraphSchedule() function to enable that one fragment is on, and do the consequtive kinfu.

NEXT:

We need to take out the integration part.
Registration only!
toggle --fragment_distance